refactor config

+++ pkg/ratelimit/limiter.go
@@ -7,13 +7,13 @@ import (
 
 type Limiter struct {
 	mu       sync.Mutex
-	rate     float64
-	capacity int64
+	rate     int
+	capacity int
 	tokens   float64
 	last     time.Time
 }
 
-func NewLimiter(rate float64, capacity int64) *Limiter {
+func NewLimiter(rate int, capacity int) *Limiter {
 	return &Limiter{
 		rate:     rate,
 		capacity: capacity,
@@ -30,7 +30,7 @@ func (l *Limiter) Allow() bool {
 	elapsed := now.Sub(l.last)
 	l.last = now
 
-	l.tokens += float64(elapsed) * l.rate / float64(time.Second)
+	l.tokens += float64(elapsed) * float64(l.rate) / float64(time.Second)
 	if l.tokens > float64(l.capacity) {
 		l.tokens = float64(l.capacity)
 	}
